Reading what a module wrote is not evaluating what it does
Two things were in `ExampleVerifier`. One evaluates a source's rows and reports what did not hold.
The other reads a module's written statements against each other and reports where two of them
answer differently — a row of an `example` says what a behavior will owe; a `fake` table and a
`with` say what stands in for it while some *other* behavior's row runs, and where those answers
differ the model contradicts itself whichever one is right. That is readable from the text alone,
with no behavior applied and no result to compare against.

They answer different questions for different keys, and neither reads the other's state. So
`ExampleStatements` is its own class: `disagreements`, the reading of each statement within its own
share of the budget, and the comparison that holds two written answers against each other.

`fakeTables` goes with it, because a written table is a written statement. It is built because it
is written — a module can write a table that nothing depends on and no row records, and it still
states what it states — and it was already the only other reader that needed a statement's own
budget. With both gone, `within` and `Read` leave the verifier entirely.

The fake's table is what the two still share, and it is shared on purpose. A second reading of one
table is what E1919 exists to catch, so there must not be a second reading to drift from the first.
`standins` builds the whole table or none of it, and stays one method that both the proxy a row runs
against and the reading that holds the table to the recorded rows call. `standingIn` is gone as a
static utility and is now `Standins.answering` — the table's own operation. Sharing a rule by asking
the thing it is about is harder to undo than sharing it by both calling the same helper, and adding
a second dispatch now means writing it somewhere it plainly does not belong. It still compares by
`souther.runtime.Values.equal` and not Java's, so a row written `1.0m` is the row a call arriving
with `1m` wants.

`Statement`, `Disagreement`, `TimedOutFake` and `Readings` move with what produces them, so `Output`
and the disagreement test name the class that answers rather than the one that used to. Nothing they
expect changed.

It is 1052, 849 and 1000 lines, from the 2709 this began as.

Refs #316
